pipeline/requirements.txt
pytest
pylint
bs4
requests
python-dotenv